// client/src/Models/followerModel.ts
// class FollowerModel {
//     public userUuid: string;
//     public vacationId: number;

//     public constructor (follower: FollowerModel) {
//         this.userUuid = follower.userUuid;
//         this.vacationId = follower.vacationId;
//     }
// }

// export default FollowerModel;

export interface FollowerModel {
    userUuid: string;
    vacationId: number;
}